Refactor folder structure to make it possible to test and deploy sensible code without significant rewrites
