+++ connection/__init__.py
@@ -1,0 +1,9 @@
+"""
+This modules holds the connections to other data sources
+such as SQLAlchemy and Marshmallow
+"""
+
+from .models import init_database
+from .schemas import init_marshmallow
+
+__all__ = [init_marshmallow, init_database]
