+++ create_and_mint_case_using_stores.py
@@ -13,8 +13,10 @@ Then...
 by retrieving casefields from the Tank and Fluid stores.
 """
 
-from sqlalchemy_classes import (init_database, db, Case, CaseField,
-                                JobParameterTemplate, JobParameter, Job)
+from connection import init_database
+from connection.models import (db, Case, CaseField,
+                               JobParameterTemplate, JobParameter, Job)
+
 from create_case_store import make_tank_store, make_fluid_store
 from create_mint_store import make_mint_store
